refactor of input.
