+++ front/website/vite.config.ts
@@ -15,6 +15,7 @@ export default defineConfig({
       input: {
         main: resolve(__dirname, 'index.html'),
         orm: resolve(__dirname, 'orm.html'),
+        test: resolve(__dirname, 'test.html'),
         cv: resolve(__dirname, 'cv.html'),
         lsys: resolve(__dirname, 'projects/l-system.html'),
         calc: resolve(__dirname, 'projects/intrest.html'),
